requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
daphne==4.1.0
Django==4.2.10
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-redis==5.4.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
flower==2.0.1
humanize==4.9.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
iniconfig==2.0.0
kombu==5.3.5
msgpack==1.0.8
numpy==1.26.3
packaging==23.2
pandas==2.2.0
pluggy==1.3.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydocstyle==6.3.0
pyOpenSSL==24.0.0
pytest==7.4.4
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==24.1.0
setuptools==69.0.3
six==1.16.0
snowballstemmer==2.2.0
sqlparse==0.4.4
tornado==6.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.2
